refactor: refactor and fix sliders
