+++ culture-portal/src/assets/i18n/en.json
@@ -33,6 +33,11 @@
   "our": "Our ",
   "team": "team",
   "timeline": "Biography",
+  "gallery":"Photo Gallery",
+  "slider": ["https://images.wallpaperscraft.ru/image/fotoapparat_obektiv_optika_141755_1920x1200.jpg",
+    "https://images.wallpaperscraft.ru/image/fotoapparat_obektiv_sumka_139250_1920x1200.jpg",
+    "https://www.vladtime.ru/uploads/posts/2018-02/1518885831_fuji-x100t-digital-camera_1920x1200_wallpaper.jpg"
+  ],
   "authors": [{
       "id": 1,
       "name": "Dashkevich Lev Urbanovich",
