+++ culture-portal/src/components/gallery/gallery.jsx
@@ -1,7 +1,6 @@
 import React, { Fragment } from 'react';
-import i18n from 'i18next';
+import { useTranslation } from 'react-i18next';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
-import './gallery.css';
 import { uniqueId } from 'lodash';
 import Photo from '../image';
 
@@ -9,27 +8,13 @@ import Photo from '../image';
 const { Carousel } = require('react-responsive-carousel');
 
 const Gallery = (props) => {
+  const { t } = useTranslation();
   const { data } = props;
-  let photoGallery;
-  switch (i18n.language) {
-    case 'ru':
-      photoGallery = 'Фотогалерея';
-      break;
-    case 'by':
-      photoGallery = 'Фотагалерэя';
-      break;
-    case 'en':
-      photoGallery = 'Photo Gallery';
-      break;
-    default:
-      photoGallery = 'Фотогалерея';
-      break;
-  }
+
   return (
     <Fragment>
-      <h3 className="photo_gallery">{photoGallery}</h3>
+      <h3 className="photo_gallery">{t('gallery')}</h3>
       <Carousel
-        // width="700px"
         dynamicHeight
         transitionTime={1000}
         emulateTouch
@@ -44,7 +29,6 @@ const Gallery = (props) => {
           </div>
         ))
         }
-        {/* <img src="https://www.znyata.com/images/stories/line/anempodystov/michal-10.jpg" alt="22" /> */}
       </Carousel>
     </Fragment>
   );
