+++ culture-portal/src/components/gallery/index.js
@@ -1,0 +1,5 @@
+import Gallery from './gallery';
+
+import './gallery.css';
+
+export default Gallery;
