+++ culture-portal/src/components/main-slider/index.js
@@ -1,0 +1,3 @@
+import MainSlider from './main-slider';
+
+export default MainSlider;
