+++ culture-portal/src/pages/index.js
@@ -4,15 +4,15 @@ import Layout from '../components/layout';
 import SEO from '../components/seo';
 import Footer from '../components/footer';
 import AuthorOfTheDay from '../components/author-of-the-day';
-import MainSlider from '../components/main-slider/main-slider';
+import MainSlider from '../components/main-slider';
 
 import '../components/translate/i18n';
 import Description from '../components/portal-description/description';
 
 const IndexPage = () => (
   <Layout>
-    <SEO title="Home" />
     <MainSlider />
+    <SEO title="Home" />
     <Description />
     <AuthorOfTheDay />
     <Footer />
